framebuffer: Reduce size of local pixel counter to uint32_t
As in the previous commit this might improve performance for some setups

+++ framebuffer.h
@@ -34,7 +34,7 @@ struct fb {
 	unsigned numa_node;
 	struct llist_entry list;
 #ifdef FEATURE_PIXEL_COUNT
-	uint64_t pixel_count;
+	uint32_t pixel_count;
 #endif
 };
